src/main/webapp/app/entities/fluxo-modelo/delete/fluxo-modelo-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { provideHttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { FluxoModeloService } from '../service/fluxo-modelo.service';

import { FluxoModeloDeleteDialogComponent } from './fluxo-modelo-delete-dialog.component';

describe('FluxoModelo Management Delete Component', () => {
  let comp: FluxoModeloDeleteDialogComponent;
  let fixture: ComponentFixture<FluxoModeloDeleteDialogComponent>;
  let service: FluxoModeloService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [FluxoModeloDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(FluxoModeloDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(FluxoModeloDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(FluxoModeloService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
